Issue #28: Add the possibility to restrict visible dropdowns and periods

+++ src/fields/Period.tsx
@@ -1,48 +1,74 @@
 import React, { useCallback, useMemo } from 'react'
-import Select from 'antd/lib/select'
+import Select, { BaseOptionType } from 'antd/lib/select'
 
 import { PeriodProps } from '../types'
 import { DEFAULT_LOCALE_EN } from '../locale'
 import { classNames } from '../utils'
 
 export default function Period(props: PeriodProps) {
-  const { value, setValue, locale, className, disabled, readOnly, shortcuts } =
-    props
-  let options = [
-    {
+  const {
+    value,
+    setValue,
+    locale,
+    className,
+    disabled,
+    readOnly,
+    shortcuts,
+    allowedPeriods,
+  } = props
+  const options: BaseOptionType[] = []
+
+  if (allowedPeriods.includes('year')) {
+    options.push({
       value: 'year',
       label: locale.yearOption || DEFAULT_LOCALE_EN.yearOption,
-    },
-    {
+    })
+  }
+
+  if (allowedPeriods.includes('month')) {
+    options.push({
       value: 'month',
       label: locale.monthOption || DEFAULT_LOCALE_EN.monthOption,
-    },
-    {
+    })
+  }
+
+  if (allowedPeriods.includes('week')) {
+    options.push({
       value: 'week',
       label: locale.weekOption || DEFAULT_LOCALE_EN.weekOption,
-    },
-    {
+    })
+  }
+
+  if (allowedPeriods.includes('day')) {
+    options.push({
       value: 'day',
       label: locale.dayOption || DEFAULT_LOCALE_EN.dayOption,
-    },
-    {
+    })
+  }
+
+  if (allowedPeriods.includes('hour')) {
+    options.push({
       value: 'hour',
       label: locale.hourOption || DEFAULT_LOCALE_EN.hourOption,
-    },
-    {
+    })
+  }
+
+  if (allowedPeriods.includes('minute')) {
+    options.push({
       value: 'minute',
       label: locale.minuteOption || DEFAULT_LOCALE_EN.minuteOption,
-    },
-  ]
+    })
+  }
 
-  if (shortcuts && (shortcuts === true || shortcuts.includes('@reboot'))) {
-    options = [
-      ...options,
-      {
-        value: 'reboot',
-        label: locale.rebootOption || DEFAULT_LOCALE_EN.rebootOption,
-      },
-    ]
+  if (
+    allowedPeriods.includes('reboot') &&
+    shortcuts &&
+    (shortcuts === true || shortcuts.includes('@reboot'))
+  ) {
+    options.push({
+      value: 'reboot',
+      label: locale.rebootOption || DEFAULT_LOCALE_EN.rebootOption,
+    })
   }
 
   const handleChange = useCallback(
